requirements.txt
appdirs==1.4.4
awscrt==0.9.15
awsiotsdk==1.5.3
boto3==1.21.4
botocore==1.24.10
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
pbr==5.5.1
pyfiglet==0.8.post1
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.5.0
six==1.15.0
stevedore==3.3.0
typing-extensions==3.10.0.0
urllib3==1.26.2
virtualenv==20.3.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
zipp==3.4.0
